Trimmed down application: Remove GUI as it's not going to be maintained in this fork.
