--- IlMerge.bat
@@ -1,3 +1,0 @@
-ilmerge /target:winexe /out:XSDDiagram.exe bin\Release\XSDDiagram.exe bin\Release\XSDDiagrams.dll
-copy XSDDiagramConsole\bin\Release\XSDDiagramConsole.exe .
-pause
