+++ XSDDiagram2010.csproj
@@ -84,92 +84,7 @@
     <Reference Include="System.Drawing" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AboutForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="AboutForm.Designer.cs">
-      <DependentUpon>AboutForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="DiagramControl.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="DiagramControl.Designer.cs">
-      <DependentUpon>DiagramControl.cs</DependentUpon>
-    </Compile>
-    <Compile Include="DiagramControlContainer.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="ElementsForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="ElementsForm.Designer.cs">
-      <DependentUpon>ElementsForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="ErrorReportForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="ErrorReportForm.Designer.cs">
-      <DependentUpon>ErrorReportForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="FileShellExtension.cs" />
-    <Compile Include="LoginPromptForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="LoginPromptForm.Designer.cs">
-      <DependentUpon>LoginPromptForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="MainForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="MainForm.Designer.cs">
-      <DependentUpon>MainForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="MRUManager.cs" />
-    <Compile Include="OpenURLForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="OpenURLForm.Designer.cs">
-      <DependentUpon>OpenURLForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Options.cs" />
-    <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Settings.cs" />
-    <Compile Include="SettingsForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="SettingsForm.Designer.cs">
-      <DependentUpon>SettingsForm.cs</DependentUpon>
-    </Compile>
-    <EmbeddedResource Include="AboutForm.resx">
-      <SubType>Designer</SubType>
-      <DependentUpon>AboutForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="DiagramControl.resx">
-      <DependentUpon>DiagramControl.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="DiagramControlContainer.resx">
-      <SubType>Designer</SubType>
-      <DependentUpon>DiagramControlContainer.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="ElementsForm.resx">
-      <SubType>Designer</SubType>
-      <DependentUpon>ElementsForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="ErrorReportForm.resx">
-      <SubType>Designer</SubType>
-      <DependentUpon>ErrorReportForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="LoginPromptForm.resx">
-      <DependentUpon>LoginPromptForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="MainForm.resx">
-      <SubType>Designer</SubType>
-      <DependentUpon>MainForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="OpenURLForm.resx">
-      <DependentUpon>OpenURLForm.cs</DependentUpon>
-    </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
@@ -180,9 +95,6 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
-    <EmbeddedResource Include="SettingsForm.resx">
-      <DependentUpon>SettingsForm.cs</DependentUpon>
-    </EmbeddedResource>
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
