+++ XSDDiagram2012.sln
@@ -1,8 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual C# Express 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XSDDiagram2010", "XSDDiagram2010.csproj", "{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}"
-EndProject
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32112.339
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XSDDiagramConsole2010", "XSDDiagramConsole\XSDDiagramConsole2010.csproj", "{1D732DFD-D631-4BFB-8057-C34B9327A08F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XSDDiagrams2010", "XSDDiagrams\XSDDiagrams2010.csproj", "{6D3DACE7-D51B-4259-883A-6899BF4D52E2}"
@@ -15,14 +15,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}.Debug|x86.ActiveCfg = Debug|x86
-		{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}.Debug|x86.Build.0 = Debug|x86
-		{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}.Release|x86.ActiveCfg = Release|x86
-		{B212CAFE-B4C8-44CB-B82D-AA32B55D2719}.Release|x86.Build.0 = Release|x86
 		{1D732DFD-D631-4BFB-8057-C34B9327A08F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1D732DFD-D631-4BFB-8057-C34B9327A08F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D732DFD-D631-4BFB-8057-C34B9327A08F}.Debug|x86.ActiveCfg = Debug|x86
@@ -42,6 +34,9 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {082865D1-2FCB-4EDF-8CB5-2C99522B2F96}
 	EndGlobalSection
 	GlobalSection(MonoDevelopProperties) = preSolution
 		StartupItem = XSDDiagram.csproj
